feat: multi_process test, refactor
